HashTable baseline and basic improvements
This is preliminary to the work to the Hash Table refactor.  It fixes
bugs and infelicities in the current Hash Table interface before the
implementation of latching hash table operations and better memory
handling.  Improvements include:

  * Use of void* rather than caddr_t in the hash table buffer
    descriptors.  This removes the need to litter the code with casts
    (and the attendant typing problems.)

  * Update the documentation to match the code in HashTable.c.

  * Fix a few lock and memory leaks.

  * Make the index into the array of Red-Black trees a uint32_t and
    the hash in a red-black tree a uint64_t.  This allows one to hash
    things like clientids and pointers using a modulus for one and the
    identity function for the other.

  * Use POSIX read-write locks.  There was talk of doing using them on
    the mailing list, and, more importantly, they have a single
    'unlock' operation rather than an 'unlock-read' or 'unlock-write'
    operation.  This is important for the 'latched' hash access to
    come.

+++ src/IdMapper/idmapper_cache.c
@@ -82,8 +82,8 @@ hash_table_t *ht_uidgid;
  * @see HashTable_Init
  *
  */
-unsigned long idmapper_value_hash_func(hash_parameter_t * p_hparam,
-                                       hash_buffer_t * buffclef)
+uint32_t idmapper_value_hash_func(hash_parameter_t * p_hparam,
+                                  hash_buffer_t * buffclef)
 {
   unsigned int sum = 0;
   unsigned int i = 0;
@@ -97,7 +97,7 @@ unsigned long idmapper_value_hash_func(hash_parameter_t * p_hparam,
 }                               /*  ip_name_value_hash_func */
 
 
-unsigned long namemapper_value_hash_func(hash_parameter_t * p_hparam,
+uint32_t namemapper_value_hash_func(hash_parameter_t * p_hparam,
                                          hash_buffer_t * buffclef)
 {
   return ((unsigned long)(buffclef->pdata) % p_hparam->index_size);
@@ -117,8 +117,8 @@ unsigned long namemapper_value_hash_func(hash_parameter_t * p_hparam,
  * @see HashTable_Init
  *
  */
-unsigned long idmapper_rbt_hash_func(hash_parameter_t * p_hparam,
-                                     hash_buffer_t * buffclef)
+uint64_t idmapper_rbt_hash_func(hash_parameter_t * p_hparam,
+                                hash_buffer_t * buffclef)
 {
   unsigned int result;
 
@@ -129,8 +129,8 @@ unsigned long idmapper_rbt_hash_func(hash_parameter_t * p_hparam,
   return (unsigned long)result;
 }                               /* ip_name_rbt_hash_func */
 
-unsigned long namemapper_rbt_hash_func(hash_parameter_t * p_hparam,
-                                       hash_buffer_t * buffclef)
+uint64_t namemapper_rbt_hash_func(hash_parameter_t * p_hparam,
+                                  hash_buffer_t * buffclef)
 {
   return (unsigned long)(buffclef->pdata);
 }
