+++ src/include/nfs_ip_stats.h
@@ -72,17 +72,17 @@ int nfs_ip_name_populate(char *path);
 int display_ip_name_key(hash_buffer_t * pbuff, char *str);
 int display_ip_name_val(hash_buffer_t * pbuff, char *str);
 int compare_ip_name(hash_buffer_t * buff1, hash_buffer_t * buff2);
-unsigned long int ip_name_rbt_hash_func(hash_parameter_t * p_hparam,
-                                        hash_buffer_t * buffclef);
-unsigned long int ip_name_value_hash_func(hash_parameter_t * p_hparam,
-                                          hash_buffer_t * buffclef);
+uint64_t ip_name_rbt_hash_func(hash_parameter_t * p_hparam,
+                               hash_buffer_t * buffclef);
+uint32_t ip_name_value_hash_func(hash_parameter_t * p_hparam,
+                                 hash_buffer_t * buffclef);
 
 int display_ip_stats_key(hash_buffer_t * pbuff, char *str);
 int display_ip_stats_val(hash_buffer_t * pbuff, char *str);
 int compare_ip_stats(hash_buffer_t * buff1, hash_buffer_t * buff2);
-unsigned long int ip_stats_rbt_hash_func(hash_parameter_t * p_hparam,
+uint64_t ip_stats_rbt_hash_func(hash_parameter_t * p_hparam,
                                          hash_buffer_t * buffclef);
-unsigned long int ip_stats_value_hash_func(hash_parameter_t * p_hparam,
-                                           hash_buffer_t * buffclef);
+uint32_t ip_stats_value_hash_func(hash_parameter_t * p_hparam,
+                                  hash_buffer_t * buffclef);
 
 #endif
