+++ src/support/nfs_ip_stats.c
@@ -81,7 +81,7 @@
  * @see HashTable_Init
  *
  */
-unsigned long int ip_stats_value_hash_func(hash_parameter_t * p_hparam,
+uint32_t ip_stats_value_hash_func(hash_parameter_t * p_hparam,
                                            hash_buffer_t * buffclef)
 {
   return hash_sockaddr((sockaddr_t *)buffclef->pdata, IGNORE_PORT) % p_hparam->index_size;
@@ -102,7 +102,7 @@ unsigned long int ip_stats_value_hash_func(hash_parameter_t * p_hparam,
  * @see HashTable_Init
  *
  */
-unsigned long int ip_stats_rbt_hash_func(hash_parameter_t * p_hparam,
+uint64_t ip_stats_rbt_hash_func(hash_parameter_t * p_hparam,
                                          hash_buffer_t * buffclef)
 {
   return hash_sockaddr((sockaddr_t *)buffclef->pdata, IGNORE_PORT);
